Add a create task button to the navbar
