+++ app/components/Sidebar/Sidebar.tsx
@@ -3,6 +3,8 @@ import React from "react";
 import menu from "../../utils/menu";
 import Link from "next/link";
 import { usePathname, useRouter } from "next/navigation";
+import { plus } from "@/app/utils/Icons";
+import { useGlobalState } from "@/app/context/globalProvider";
 
 interface props {
 	isMobile: boolean;
@@ -14,6 +16,27 @@ function Sidebar({ isMobile }: props) {
 
 	const handleClick = (link: string) => {
 		router.push(link);
+	};
+
+	const { createTask } = useGlobalState();
+
+	const makeNavbarItems = (items: Array<any>) => {
+		return items.map((item) => {
+			const link = item.link;
+			return (
+				<li
+					key={item.id}
+					className={`nav-item ${pathname === link ? "active" : ""}`}
+					onClick={() => {
+						handleClick(link);
+					}}
+				>
+					<Link href={link}>
+						{item.icon} {!isMobile && item.title}
+					</Link>
+				</li>
+			);
+		});
 	};
 
 	return (
@@ -28,24 +51,23 @@ function Sidebar({ isMobile }: props) {
 				</ul>
 			)}
 			<ul className="nav-items">
-				{menu.map((item) => {
-					const link = item.link;
-					return (
-						<li
-							key={item.id}
-							className={`nav-item ${
-								pathname === link ? "active" : ""
-							}`}
-							onClick={() => {
-								handleClick(link);
-							}}
-						>
-							<Link href={link}>
-								{item.icon} {!isMobile && item.title}
-							</Link>
+				{makeNavbarItems(menu.slice(0, 2))}
+				{isMobile && (
+					<button className="btn-rounded" onClick={createTask}>
+						{plus}
+					</button>
+				)}
+				{makeNavbarItems(menu.slice(2))}
+
+				{!isMobile && (
+						<li className="nav-item">
+							<Link href="/about">About</Link>
 						</li>
-					);
-				})}
+					) && (
+						<button className="btn-rounded" onClick={createTask}>
+							{plus}
+						</button>
+					)}
 			</ul>
 		</nav>
 	);
