+++ styles/globals.scss
@@ -68,30 +68,6 @@ footer {
 		width: 0.5rem;
 	}
 
-	.btn-rounded {
-		position: fixed;
-		top: 4.9rem;
-		right: 5.1rem;
-		width: 3rem;
-		height: 3rem;
-		border-radius: 50%;
-		opacity: 1;
-
-		border: 2px solid $default-border-color2;
-		box-shadow: 0 3px 15px rgba(0, 0, 0, 0.3);
-		color: $default-color-grey0;
-		font-size: 1.4rem;
-
-		display: flex;
-		align-items: center;
-		justify-content: center;
-
-		@media screen and (max-width: 768px) {
-			top: 3rem;
-			right: 3.5rem;
-		}
-	}
-
 	.tasks {
 		margin: 2rem 0;
 
@@ -250,6 +226,14 @@ footer {
 				}
 			}
 		}
+
+		.btn-rounded {
+			height: 110%;
+			aspect-ratio: 1;
+            z-index: 100;
+			border-radius: 50%;
+			margin: 0 1rem;
+		}
 	}
 
 	.nav-item {
@@ -319,6 +303,16 @@ footer {
 
 	.active::before {
 		height: 0.3rem;
+	}
+
+	.btn-rounded {
+		box-shadow: 0 3px 15px rgba(0, 0, 0, 0.3);
+		color: $default-color-grey0;
+		font-size: 1.4rem;
+
+		display: flex;
+		align-items: center;
+		justify-content: center;
 	}
 
 	> button {
